refactor: Node Entity 추가
