+++ apps/frontend/src/entities/node/index.ts
@@ -1,0 +1,8 @@
+export {
+  type Node,
+  type NoteNodeData,
+  type NoteNodeType,
+} from "./model/nodeTypes";
+
+export { MemoizedGroupNode } from "./ui/GroupNode";
+export { NoteNode } from "./ui/NoteNode";
