+++ apps/frontend/src/entities/node/model/nodeTypes.ts
@@ -1,0 +1,13 @@
+import { type Node as FlowNode } from "@xyflow/react";
+
+export interface Node extends FlowNode {
+  isHolding: boolean;
+}
+
+export type NoteNodeData = {
+  title: string;
+  id: number;
+  emoji: string;
+};
+
+export type NoteNodeType = FlowNode<NoteNodeData, "note">;
