--- apps/frontend/src/features/canvas/ui/Node/index.tsx
@@ -1,2 +1,0 @@
-export { GroupNode } from "./GroupNode";
-export { NoteNode } from "./NoteNode";
